requirements.txt
fastapi
uvicorn
google-auth
google-api-python-client
google-cloud-pubsub
google-cloud-secret-manager
requests